Add ProviderFuture method for sync on local operations
Add method 'hotSync' which can be optimized for calls that
are expected to be local only or very quick to complete.
The other 'sync' methods can then be simplified to run in
a wait / notify pattern as those operations are expected
to be reound trips to the remote and therefore are not a
candidate for optimized completion waits.