Refactor Anonymous fallback producer to cache and close as needed
Because we want to avoid spawning many links without bound the cache
should be fixed sized and limit new producers until old one's are
closed.  Also we want to ensure that we manage producer lifetime
efficiently so we run a sweeper to clean on producers from the cache
instead of schedule close on each send.